article-summarizer-extension/backend/requirements.txt
flask==2.3.3
newspaper3k==0.2.8
flask-cors==4.0.0
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
requests==2.31.0